Refactor the Alphabet & Hsm types into separate modules.
This is easy / possible now that the Hsm type implements RngCore.
